tests/{a32,a64}_unicorn: Avoid use of deprecated fmt API

+++ tests/unicorn_emu/a32_unicorn.cpp
@@ -12,12 +12,12 @@
 
 #include "../A32/testenv.h"
 
-#define CHECKED(expr)                                                               \
-    do {                                                                            \
-        if (auto cerr_ = (expr)) {                                                  \
-            ASSERT_MSG(false, "Call " #expr " failed with error: {} ({})\n", cerr_, \
-                       uc_strerror(cerr_));                                         \
-        }                                                                           \
+#define CHECKED(expr)                                                                                    \
+    do {                                                                                                 \
+        if (auto cerr_ = (expr)) {                                                                       \
+            ASSERT_MSG(false, "Call " #expr " failed with error: {} ({})\n", static_cast<size_t>(cerr_), \
+                       uc_strerror(cerr_));                                                              \
+        }                                                                                                \
     } while (0)
 
 constexpr u32 BEGIN_ADDRESS = 0;
@@ -52,7 +52,7 @@ void A32Unicorn<TestEnvironment>::Run() {
             return;
         }
         if (auto cerr_ = uc_emu_start(uc, pc, END_ADDRESS, 0, 1)) {
-            fmt::print("uc_emu_start failed @ {:08x} (code = {:08x}) with error {} ({})", pc, *testenv.MemoryReadCode(pc), cerr_, uc_strerror(cerr_));
+            fmt::print("uc_emu_start failed @ {:08x} (code = {:08x}) with error {} ({})", pc, *testenv.MemoryReadCode(pc), static_cast<size_t>(cerr_), uc_strerror(cerr_));
             throw "A32Unicorn::Run() failure";
         }
         testenv.ticks_left--;
