+++ tests/unicorn_emu/a64_unicorn.cpp
@@ -7,12 +7,12 @@
 
 #include <mcl/assert.hpp>
 
-#define CHECKED(expr)                                                               \
-    do {                                                                            \
-        if (auto cerr_ = (expr)) {                                                  \
-            ASSERT_MSG(false, "Call " #expr " failed with error: {} ({})\n", cerr_, \
-                       uc_strerror(cerr_));                                         \
-        }                                                                           \
+#define CHECKED(expr)                                                                                    \
+    do {                                                                                                 \
+        if (auto cerr_ = (expr)) {                                                                       \
+            ASSERT_MSG(false, "Call " #expr " failed with error: {} ({})\n", static_cast<size_t>(cerr_), \
+                       uc_strerror(cerr_));                                                              \
+        }                                                                                                \
     } while (0)
 
 constexpr u64 BEGIN_ADDRESS = 0;
